requirements.txt
fire==0.4.0
gin-config==0.5.0
matplotlib==3.5.1
numpy ==1.22.0
pandas==1.4.2
scikit-learn==1.0.2
scipy==1.7.3
tensorboard==2.8.0
torch==1.11.0
torchvision==0.12.0   # CUDA 11.3 is tested
#torch==2.1.2+cu121
#torchvision==0.16.2  #CUDA 12.1 is tested
tqdm==4.62.3
einops==0.4.1
protobuf==3.20.0